fix int casting in brush

+++ ToolD3DX/cBrush.cpp
@@ -39,25 +39,25 @@ void cBrush::GetRect()
 {
     int nSize = (g_pMapDataManager->GetMapSize() + 1) * 64;
 
-    int nMinX = (m_pPick.x - m_fBrushRadius);
+    int nMinX = (int)(m_pPick.x - m_fBrushRadius);
     if (nMinX < 0)
     {
         nMinX = 0;
     }
 
-    int nMaxX = (m_pPick.x + m_fBrushRadius) * nSize + 1;
+    int nMaxX = (int)((m_pPick.x + m_fBrushRadius) * nSize + 1);
     if (nMaxX > nSize)
     {
         nMaxX = nSize;
     }
 
-    int nMinZ = (m_pPick.z - m_fBrushRadius) * nSize;
+    int nMinZ = (int)((m_pPick.z - m_fBrushRadius) * nSize);
     if (nMinZ < 0)
     {
         nMinZ = 0;
     }
 
-    int nMaxZ = (m_pPick.z + m_fBrushRadius) * nSize + 1;
+    int nMaxZ = (int)((m_pPick.z + m_fBrushRadius) * nSize + 1);
     if (nMaxZ > nSize)
     {
         nMaxZ = nSize;
@@ -68,25 +68,25 @@ void cBrush::GetRect()
     m_nNearMinX = nMinX;
     m_nNearMinZ = nMinZ;
 
-    nMinX = ((m_pPick.x - m_fSprayRadius) * nSize);
+    nMinX = (int)((m_pPick.x - m_fSprayRadius) * nSize);
     if (nMinX < 0)
     {
         nMinX = 0;
     }
 
-    nMaxX = ((m_pPick.x + m_fSprayRadius) * nSize + 1);
+    nMaxX = (int)((m_pPick.x + m_fSprayRadius) * nSize + 1);
     if (nMaxX > nSize)
     {
         nMaxX = nSize;
     }
 
-    nMinZ = ((m_pPick.z - m_fSprayRadius) * nSize);
+    nMinZ = (int)((m_pPick.z - m_fSprayRadius) * nSize);
     if (nMinZ < 0)
     {
         nMinZ = 0;
     }
 
-    nMaxZ = ((m_pPick.z + m_fSprayRadius) * nSize + 1);
+    nMaxZ = (int)((m_pPick.z + m_fSprayRadius) * nSize + 1);
     if (nMaxZ > nSize)
     {
         nMaxZ = nSize;
@@ -96,5 +96,4 @@ void cBrush::GetRect()
     m_nFarMaxZ = nMaxZ;
     m_nFarMinX = nMinX;
     m_nFarMinZ = nMinZ;
-    
 }
